Lower devise min password length for ALPHA.
There are some issues with how the registration form works right now when
handling server side errors, and so if you entered a password that was too short
it wasn't clear that's what the problem is. As we're still in early alpha and
a lot might change (including the reg form) I'm temporarily lowering the minimum
password length as a workaround.