Fixes #10. Fix bug where the session user might be disposed too soon.
The process of registering consists of creating the user (via API), then setting
a user into Chaplin.mediator.user. Previously we were setting the registered
user object directly, which broke because when we left the registration page
we would dispose that user, clearing it's properties. The fix is to always create
a new user to be the "session user", copying attributes from the user object
given. This way we know our "session user" isn't attached to any other views
or controllers.